feat(rdb-panel): estandariza panel /rdb con cards + mueve home operativa a /rdb/home
Beto pidió igualar el panel /rdb al patrón de /dilesa (cards de
módulos agrupados por sección, filtrados por permisos del usuario
efectivo) y mover el dashboard operativo actual a su URL propia para
controlarlo con permisos por rol.

Cambios:

- app/rdb/page.tsx (reescrito) — panel de cards con 6 secciones
  (Operativa, Administración, RRHH, Compras, Inventario, Operaciones).
  Mismo patrón que app/dilesa/page.tsx: usePermissions + canAccessModulo
  + ROUTE_TO_MODULE para filtrar por rol del usuario efectivo. Solo
  RequireAccess empresa="rdb"; el filtrado de cards lo hace el panel.

- app/rdb/home/page.tsx (movido desde app/rdb/page.tsx) — el dashboard
  operativo de 823 líneas ahora vive aquí. Mantiene RequireAccess
  empresa="rdb" modulo="rdb.home" para gatear acceso por rol.

- lib/permissions.ts: ROUTE_TO_MODULE mapea ahora '/rdb/home' →
  'rdb.home' (en vez de '/rdb' → 'rdb.home'). El slug DB no cambia,
  solo la URL frontend que lo dispara.

- components/app-shell/nav-config.ts: agrega sección "Operativa" como
  sections[0] de RDB con un solo item Home en /rdb/home. Aparece como
  primer renglón al expandir RDB en el sidebar.

- components/app-shell/__tests__/nav-config.test.ts: actualiza el test
  flattenNavChildren para reflejar que el primer item de RDB ahora es
  /rdb/home (no /rdb/admin/tasks).

Sin migración DB: el slug rdb.home ya existía en core.modulos. Solo
cambia la URL UI; los permisos por rol se mantienen.

Bookmarks viejos a /rdb (que antes era home operativa) ahora aterrizan
en el panel de cards. Desde ahí navegan a /rdb/home en un click.
